requirements.txt
streamlit
pandas
pyarrow
beautifulsoup4
lxml
tldextract
playwright
langchain
langchain-ollama
langchain-core
pypdf
python-docx
uvicorn
fastapi
pydantic
httpx
openai>=1.0.0
slowapi==0.1.9
limits>=3.0
